refactor(storage): Split redb/tests.rs and test_suite/iterators.rs into focused test modules
Split oversized test files for context-efficient agent navigation:
- redb/tests.rs (1,486 lines) → tests/ directory with 6 focused modules
- test_suite/iterators.rs (948 lines) → 4 focused iterator test files
