+++ crates/storage/src/backends/test_suite/mod.rs
@@ -23,12 +23,18 @@ mod basic_ops;
 mod callbacks;
 mod concurrency;
 mod dropable;
-mod iterators;
+mod iterator_basic;
+mod iterator_edge_cases;
+mod iterator_reverse;
+mod iterator_seek;
 mod macros;
 
 pub use basic_ops::*;
 pub use callbacks::*;
 pub use concurrency::*;
 pub use dropable::*;
-pub use iterators::*;
+pub use iterator_basic::*;
+pub use iterator_edge_cases::*;
+pub use iterator_reverse::*;
+pub use iterator_seek::*;
 pub use macros::*;
